sql: move TestTelemetry to sqltestutils
This allows re-use in CCL.

Release justification: non-production code change.

Release note: None
