+++ pkg/sql/BUILD.bazel
@@ -517,7 +517,6 @@ go_test(
         "//pkg/security",
         "//pkg/security/securitytest",
         "//pkg/server",
-        "//pkg/server/diagnostics",
         "//pkg/server/serverpb",
         "//pkg/server/status/statuspb:statuspb_go_proto",
         "//pkg/server/telemetry",
@@ -568,7 +567,6 @@ go_test(
         "//pkg/storage",
         "//pkg/testutils",
         "//pkg/testutils/buildutil",
-        "//pkg/testutils/diagutils",
         "//pkg/testutils/jobutils",
         "//pkg/testutils/serverutils",
         "//pkg/testutils/skip",
@@ -578,7 +576,6 @@ go_test(
         "//pkg/util/bitarray",
         "//pkg/util/caller",
         "//pkg/util/cancelchecker",
-        "//pkg/util/cloudinfo",
         "//pkg/util/ctxgroup",
         "//pkg/util/encoding",
         "//pkg/util/fsm",
@@ -602,7 +599,6 @@ go_test(
         "//pkg/util/timetz",
         "//pkg/util/timeutil",
         "//pkg/util/tracing",
-        "//pkg/util/treeprinter",
         "//pkg/util/uuid",
         "@com_github_cockroachdb_apd_v2//:apd",
         "@com_github_cockroachdb_datadriven//:datadriven",
