+++ pkg/sql/sqltestutils/BUILD.bazel
@@ -5,20 +5,34 @@ go_library(
     srcs = [
         "show_create_table.go",
         "sql_test_utils.go",
+        "telemetry.go",
     ],
     importpath = "github.com/cockroachdb/cockroach/pkg/sql/sqltestutils",
     visibility = ["//visibility:public"],
     deps = [
+        "//pkg/base",
         "//pkg/config/zonepb",
         "//pkg/jobs",
         "//pkg/keys",
         "//pkg/kv",
         "//pkg/roachpb",
+        "//pkg/security",
+        "//pkg/server",
+        "//pkg/server/diagnostics",
+        "//pkg/sql",
         "//pkg/sql/catalog/catalogkv",
+        "//pkg/sql/catalog/catconstants",
         "//pkg/sql/catalog/descpb",
         "//pkg/sql/tests",
+        "//pkg/testutils/diagutils",
         "//pkg/testutils/serverutils",
+        "//pkg/testutils/skip",
+        "//pkg/testutils/sqlutils",
+        "//pkg/util/cloudinfo",
+        "//pkg/util/log",
         "//pkg/util/protoutil",
+        "//pkg/util/treeprinter",
+        "@com_github_cockroachdb_datadriven//:datadriven",
         "@com_github_cockroachdb_errors//:errors",
         "@com_github_stretchr_testify//require",
     ],
